Move kernel enum in kriging based
